refectored code
